+++ src/TC_409/TC_409.java
@@ -1,4 +1,4 @@
-package src;
+package TC_409;
 
 public class TC_409 {
 }
